requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
pillow==10.0.0
numpy==1.26.4
opencv-python==4.8.1.76
face_recognition==2.1.17